refactor: separate remote to http and firestore
