+++ lib/locator.dart
@@ -2,10 +2,10 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:mat_practice_pte/src/configs/locate/f_locate.dart';
 import 'package:mat_practice_pte/src/features/app/cubit/f_user.dart';
+import 'package:mat_practice_pte/src/networks/http/dio_http_client.dart';
+import 'package:mat_practice_pte/src/networks/http/services/definition_service.dart';
 import 'package:mat_practice_pte/src/services/audio_manger.dart';
 import 'package:mat_practice_pte/src/utils/global_variables.dart';
-import 'package:mat_practice_pte/src/utils/remote/services/definition_service.dart';
-import 'package:mat_practice_pte/src/utils/remote/services/dio_http_client.dart';
 import 'package:mat_practice_pte/src/widgets/f_app.dart';
 import 'firebase_options.dart';
 import 'src/configs/routes/app_router.dart';
