+++ lib/src/features/category/cubit/parent_category_state.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 part of 'parent_category_cubit.dart';
 
 class ParentCategoryState extends Equatable {
