+++ lib/src/features/category/ui/drawer_screen.dart
@@ -11,9 +11,9 @@ import 'package:mat_practice_pte/src/features/app/cubit/app_cubit.dart';
 import 'package:mat_practice_pte/src/features/category/cubit/drawer_cubit.dart';
 import 'package:mat_practice_pte/src/features/category/widgets/category_drawer_widget.dart';
 import 'package:mat_practice_pte/src/features/category/widgets/triangle_painter.dart';
+import 'package:mat_practice_pte/src/networks/firestore/repository/lesson_repository.dart';
+import 'package:mat_practice_pte/src/networks/models/lesson/detail_lesson.dart';
 import 'package:mat_practice_pte/src/utils/helpers/helpers.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/detail_lesson.dart';
-import 'package:mat_practice_pte/src/utils/repository/lesson_repository.dart';
 import 'package:mat_practice_pte/src/widgets/dimentions/f_padding_sizes.dart';
 import 'package:mat_practice_pte/src/widgets/dimentions/f_sizeboxs.dart';
 import 'package:mat_practice_pte/src/widgets/loading_widget.dart';
