+++ lib/src/features/word_definition/cubit/show_bottom_definition_cubit.dart
@@ -4,13 +4,13 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:mat_practice_pte/src/configs/constants/app_colors.dart';
 import 'package:mat_practice_pte/src/configs/constants/app_icons.dart';
 import 'package:mat_practice_pte/src/configs/constants/app_text_styles.dart';
+import 'package:mat_practice_pte/src/networks/fetch_resource.dart';
+import 'package:mat_practice_pte/src/networks/firestore/repository/domain_manager.dart';
+import 'package:mat_practice_pte/src/networks/models/word/m_word.dart';
+import 'package:mat_practice_pte/src/networks/models/word/meaning.dart';
+import 'package:mat_practice_pte/src/networks/models/word/phonetic.dart';
 import 'package:mat_practice_pte/src/services/audio_manger.dart';
 import 'package:mat_practice_pte/src/utils/global_variables.dart';
-import 'package:mat_practice_pte/src/utils/remote/fetch_resource.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/word/m_word.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/word/meaning.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/word/phonetic.dart';
-import 'package:mat_practice_pte/src/utils/repository/domain_manager.dart';
 import 'package:mat_practice_pte/src/utils/wrapper.dart';
 import 'package:mat_practice_pte/src/widgets/dimentions/f_padding_sizes.dart';
 import 'package:mat_practice_pte/src/widgets/dimentions/f_sizeboxs.dart';
