+++ lib/src/networks/f_result.dart
