+++ lib/src/networks/firestore/base_collection_reference.dart
