+++ lib/src/networks/firestore/repository/authenticate_repository_impl.dart
@@ -2,10 +2,11 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:mat_practice_pte/src/configs/constants/firebase_collection_names.dart';
-import 'package:mat_practice_pte/src/utils/base_collection_reference.dart';
-import 'package:mat_practice_pte/src/utils/remote/f_result.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/m_user.dart';
-import 'package:mat_practice_pte/src/utils/repository/authenticate_repository.dart';
+
+import '../../f_result.dart';
+import '../../models/user/m_user.dart';
+import '../base_collection_reference.dart';
+import 'authenticate_repository.dart';
 
 class AuthenticateRepositoryImpl extends BaseCollectionReference<MUser>
     implements AuthenticateRepository {
