+++ lib/src/networks/firestore/repository/category_repository.dart
@@ -1,5 +1,5 @@
-import 'package:mat_practice_pte/src/utils/remote/f_result.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/f_category.dart';
+import '../../f_result.dart';
+import '../../models/category/f_category.dart';
 
 abstract class CategoryRepository {
   Future<FResult<List<FCategory>>> getCategories();
