+++ lib/src/networks/firestore/repository/user_repository.dart
@@ -1,0 +1,6 @@
+import '../../f_result.dart';
+import '../../models/user/m_user.dart';
+
+abstract class UserRepository {
+  Future<FResult<MUser>> getInfoUser({required String uid});
+}
