+++ lib/src/networks/firestore/repository/user_repository_impl.dart
@@ -1,9 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:mat_practice_pte/src/configs/constants/firebase_collection_names.dart';
-import 'package:mat_practice_pte/src/utils/base_collection_reference.dart';
-import 'package:mat_practice_pte/src/utils/remote/f_result.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/m_user.dart';
-
+import '../../f_result.dart';
+import '../../models/user/m_user.dart';
+import '../base_collection_reference.dart';
 import 'user_repository.dart';
 
 class UserRepositoryImpl extends BaseCollectionReference<MUser>
