+++ lib/src/networks/http/repository/word_repository.dart
@@ -1,0 +1,6 @@
+import '../../f_result.dart';
+import '../../models/word/m_word.dart';
+
+abstract class WordRepository {
+  Future<FResult<MWord>> getDefinition(String text);
+}
