+++ lib/src/networks/http/services/definition_service.dart
@@ -1,7 +1,6 @@
 import 'package:dio/dio.dart';
-import 'package:mat_practice_pte/src/utils/remote/models/word/m_word.dart';
 import 'package:retrofit/http.dart';
-
+import '../../models/word/m_word.dart';
 part 'definition_service.g.dart';
 
 @RestApi()
