+++ lib/src/networks/models/category/f_category.dart
