+++ lib/src/networks/models/word/definition.dart
