+++ lib/src/networks/models/word/meaning.dart
