+++ lib/src/networks/models/word/phonetic.dart
