--- lib/src/utils/remote/models/word/word_dictionary.dart
@@ -1,29 +1,0 @@
-// import 'package:json_annotation/json_annotation.dart';
-
-// import 'meaning.dart';
-// import 'phonetic.dart';
-
-// part 'word_dictionary.g.dart';
-
-// @JsonSerializable()
-// class WordDictionary {
-//   @JsonKey(name: 'word')
-//   String? word;
-//   // String? phonetic;
-//   @JsonKey(name: 'phonetics')
-//   List<Phonetic>? phonetics;
-//   // String? origin;
-//   @JsonKey(name: 'meanings')
-//   List<Meaning>? meanings;
-
-//   WordDictionary({
-//     this.word,
-//     this.phonetics,
-//     this.meanings,
-//   });
-
-//   factory WordDictionary.fromJson(Map<String, dynamic> json) =>
-//       _$WordDictionaryFromJson(json);
-
-//   Map<String, dynamic> toJson() => _$WordDictionaryToJson(this);
-// }
